[refactor] Cookie -> Cookies 수정
