+++ src/main/java/webserver/http/cookie/Cookie.java
@@ -2,44 +2,66 @@ package webserver.http.cookie;
 
 import utils.StringUtils;
 
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Map;
-
-import static java.util.stream.Collectors.collectingAndThen;
-import static java.util.stream.Collectors.toMap;
-
+// TODO: 비슷한 기능을 하는 클래스가 너무 많다.
+// TODO: 중복을 제거하자.
 public class Cookie {
 
-    private static final Cookie EMPTY = new Cookie(Collections.emptyMap());
+    static final String SEPARATOR = "=";
 
-    private static final String SEPARATOR = "; ";
+    private static final int SEPARATOR_NOT_FOUND_INDEX = -1;
+    private static final int START_INDEX = 0;
 
-    private final Map<String, String> cookie;
+    private final String key;
+    private final String value;
 
-    private Cookie(final Map<String, String> cookie) {
-        this.cookie = cookie;
+    private Cookie(final String key,
+                   final String value) {
+        this.key = key;
+        this.value = value;
     }
 
-    public static Cookie of(final String rawCookie) {
+    static Cookie of(final String rawCookie) {
         if (StringUtils.isBlank(rawCookie)) {
-            return EMPTY;
+            throw new InvalidCookieException(rawCookie);
         }
 
-        return Arrays.stream(rawCookie.split(SEPARATOR))
-                .map(CookieValue::of)
-                .collect(collectingAndThen(toMap(CookieValue::getKey, CookieValue::getValue), Cookie::new));
+        final int separatorIndex = parseSeparatorIndex(rawCookie);
+
+        final String key = parseKey(rawCookie, separatorIndex);
+        final String value = parseValue(rawCookie, separatorIndex);
+
+        return new Cookie(key, value);
     }
 
-    public String getString(final String key) {
-        return cookie.get(key);
+    String getKey() {
+        return key;
     }
 
-    public boolean getBoolean(final String key) {
-        return Boolean.valueOf(getString(key));
+    String getValue() {
+        return value;
     }
 
-    public boolean isEmpty() {
-        return cookie.isEmpty();
+    private static int parseSeparatorIndex(final String rawRequestQueryValue) {
+        final int separatorIndex = rawRequestQueryValue.indexOf(SEPARATOR);
+        if (separatorIndex == SEPARATOR_NOT_FOUND_INDEX) {
+            throw new InvalidCookieException(rawRequestQueryValue);
+        }
+
+        return separatorIndex;
+    }
+
+    private static String parseKey(final String rawRequestQueryValue,
+                                   final int separatorIndex) {
+        final String key = rawRequestQueryValue.substring(START_INDEX, separatorIndex);
+        if (StringUtils.isBlank(key)) {
+            throw new InvalidCookieException(rawRequestQueryValue);
+        }
+
+        return key;
+    }
+
+    private static String parseValue(final String rawRequestQueryValue,
+                                     final int separatorIndex) {
+        return rawRequestQueryValue.substring(separatorIndex + SEPARATOR.length());
     }
 }
